The main function for parsing a fasta file

read_fasta() get the current regex

read_fasta(fasta_path = NULL, patterns = NULL, as = c("list", "data.frame"))

Arguments

fasta_path

a character string of the path to the fasta formatted file

patterns

a list, if not provided the default from regex() will be used. Note: the first element in the regex list will define the list reference name, such that with the list output, each protein can be accessed with that designation. Note: if the patterns list is missing an explicit "sequence" element, no sequence will be returned. This might be beneficial if only a few meta elements are sought.

as

a character designating the output format

Value

a list

Examples

if (FALSE) {
library(msfastar)
proteins <- read_fasta("~/Local/fasta/ecoli_UniProt.fasta")

# using a custom supplied regex list
proteins <- read_fasta(fasta_path = "~/Local/fasta/ecoli_UniProt.fasta",
                  pattern = list(
                      "accession" = "sp\\|[A-Z]",
                      "gene_name" = "(?<=GN\\=).*?(?=\\s..\\=)"
                  ))
}